requirements.txt
fastapi
uvicorn
openai
streamlit
requests
python-dotenv
